[SP-160] refactor: Chatting 관련 엔티티 BaseEntity import 수정
